+++ src/rules/matches.js
@@ -1,4 +1,3 @@
-
 export default {
     condition: field => field.hasAttribute('data-val-equalto'),
 
@@ -14,5 +13,9 @@ export default {
         }
 
         return false;
-    }
+    },
+
+    defaultMessage: 'This field does not match the {0} field.',
+
+    defaultMessageValue: field => field.getAttribute('data-val-equalto').replace('*.', '')
 };