+++ src/rules/minlength.js
@@ -1,4 +1,3 @@
-
 export default {
     condition: field => field.hasAttribute('minlength') || field.hasAttribute('data-val-minlength'),
 
@@ -10,5 +9,10 @@ export default {
 
         return field.value.trim().length >= parseInt(field.getAttribute('minlength')
             || field.getAttribute('data-val-minlength'), 10);
-    }
+    },
+
+    defaultMessage: 'This field must be at least {0} characters in length.',
+
+    defaultMessageValue: field => parseInt(field.getAttribute('minlength')
+        || field.getAttribute('data-val-minlength'), 10)
 };