+++ test/index.test.js
@@ -423,6 +423,7 @@ describe('error states', () => {
 
 });
 
+
 describe('callbacks', () => {
 
     it('should be able to add multiple callbacks to an event', () => {